feat: created getClient util
