+++ apps/agent/entrypoints/options/create-graph/GraphChat.tsx
@@ -1,0 +1,5 @@
+import type { FC } from 'react'
+
+export const GraphChat: FC = () => {
+  return <div className="flex h-full w-[50vw] flex-1"></div>
+}
