+++ apps/agent/lib/rpc/getClient.ts
@@ -1,0 +1,8 @@
+import type { AppType } from '@browseros/server'
+import { hc } from 'hono/client'
+import { getAgentServerUrl } from '../browseros/helpers'
+
+export const getClient = async () => {
+  const serverUrl = await getAgentServerUrl()
+  return hc<AppType>(serverUrl)
+}
